🏗️ Architect: Modularized Diagonal equipment
Refactored the 1,200+ line `apts/opticalequipment/diagonal.py` into a modular package structure.

Key Changes:
- Created `apts/opticalequipment/diagonal/` package.
- Isolated core `Diagonal` logic in `base.py`.
- Distributed 68 diagonal models across 22 vendor-specific modules in `vendors/`.
- Implemented dynamic loading in `__init__.py` using `pkgutil` and `importlib` to aggregate vendor data/methods into the base class.
- Maintained backward compatibility for the public API.

Benefits:
- Improves maintainability by separating data from logic.
- Reduces file size and complexity of the primary equipment classes.
- Simplifies adding new vendors in the future.
- Adheres to the Single Responsibility Principle.
